+++ lib/after-prepare.js
@@ -12,7 +12,12 @@ module.exports = function (hookArgs) {
 		release: hookArgs.prepareData.release
 	};
 
-	if (env.snapshot && shouldSnapshot(shouldSnapshotOptions)) {
+	if (env.snapshot &&
+		shouldSnapshot(shouldSnapshotOptions) &&
+		(!hookArgs.prepareData ||
+			!hookArgs.prepareData.nativePrepare ||
+			!hookArgs.prepareData.nativePrepare.skipNativePrepare)) {
+
 		installSnapshotArtefacts(hookArgs.prepareData.projectDir);
 	}
 }
