+++ lib/utils.js
@@ -1,4 +1,6 @@
 const os = require("os");
+const { dirname } = require("path");
+const { existsSync, mkdirSync } = require("fs");
 const { isAndroid } = require("../projectHelpers");
 
 function shouldSnapshot(config) {
@@ -21,9 +23,19 @@ function warn(message) {
     }
 }
 
+function ensureDirectoryExistence(filePath) {
+    var dir = dirname(filePath);
+    if (existsSync(dir)) {
+        return true;
+    }
+    ensureDirectoryExistence(dir);
+    mkdirSync(dir);
+}
+
 module.exports = {
     shouldSnapshot,
     convertToUnixPath,
     isWinOS,
-    warn
+    warn,
+    ensureDirectoryExistence
 };
